perf: hydrate the dashboard layout query so SSR stops emitting a skeleton
The consultant dashboard layout is a client component that returns
PersonalDashboardShellSkeleton instead of children whenever its queries are
loading — always true during SSR, because nothing prefetched them. Measured on
#1103: the server HTML contained no <h1>, none of the sidebar nav, and FCP sat
at ~6s while "Welcome back" arrived at 4.6s inside the RSC payload. This is the
blocker that made #1102's Suspense boundaries and #1103's StreamProvider fix
invisible; both were real, both sat underneath this.

Splits the layout the same way org-workspace already does: the client component
becomes ConsultantDashboardShell, and layout.tsx becomes a server wrapper that
seeds the query and dehydrates it.

Seeds user-details, not consultant-data. The guard is
`(...loading) && !consultantData && !userDetails`, so either one opens it, and
the consultant read carries public/private access levels, #726 plan-visibility
filtering and PII narrowing — re-deriving that for a prefetch is how #946
happened. user-details is the viewer's OWN row, so authorization is just the
session user's id, never the consultantId from the URL, which belongs to
someone else when an ADMIN/STAFF inspects.

The query moves to lib/data/user-details.ts and GET /api/user/[id] now calls it,
so the route and the prefetch cannot drift.

Not yet verified: whether <h1> and the nav actually appear in the server HTML.
That is the only thing that decides if this worked, and it needs the preview.
